Revert https://github.com/gradle/gradle/pull/22541
See https://github.com/gradle/gradle/issues/26919
