+++ subprojects/internal-integ-testing/src/main/groovy/org/gradle/integtests/fixtures/AbstractIntegrationSpec.groovy
@@ -33,7 +33,6 @@ import org.gradle.integtests.fixtures.executer.GradleExecuter
 import org.gradle.integtests.fixtures.executer.InProcessGradleExecuter
 import org.gradle.integtests.fixtures.executer.IntegrationTestBuildContext
 import org.gradle.integtests.fixtures.executer.UnderDevelopmentGradleDistribution
-import org.gradle.integtests.fixtures.logging.ProgressLoggingOutputNormalizer
 import org.gradle.integtests.fixtures.timeout.IntegrationTestTimeout
 import org.gradle.test.fixtures.dsl.GradleDsl
 import org.gradle.test.fixtures.file.CleanupTestDirectory
@@ -561,7 +560,7 @@ tmpdir is currently ${System.getProperty("java.io.tmpdir")}""")
     }
 
     String getOutput() {
-        ProgressLoggingOutputNormalizer.normalize(result.output)
+        result.output
     }
 
     String getErrorOutput() {
